qt4-libs: Deal with non-portable unused files
After extracting, remove a few files that have non-portable shell
constructs.  Both qt4-libs and qt4-tools build fine without them, so
they are obviously (slight leap) not used by the build.

(Someday, when we have a standard approach for this, this code can be
updated, but today it's important that this build.)